FIX: Rename output layer of TinyBertForSequenceTagging
The reason for changing the name of the output layer from "seq_tagger" to "classifier":
The name of the output layer of the teacher model is "classifier".
This is because of the default setting of HF Transformers.
(https://github.com/huggingface/transformers/blob/ef3cec0ca577e5950e42e8de1a2991b5dc85dfa6/src/transformers/models/bert/modeling_bert.py#L1695)
To match this as well as to utilize the HF framework for inferencing step,
the name of the output layer of our student model should be "classifier".