precompute lookups in CachedIndexGroup.search
We previously went from an entity class to an ontology to an
index all within the search method. We can precompute an
entity class -> index mapping and in doing so a.) only do the
work once for repeated lookups, and b.) move the work from search
time to startup time, which is probably preferable.