fix: KeyError in CachedIndexGroup.search
the line row_dict.pop(IDX) failed with a KeyError for results
from embedding indices because they gave results with the IDX
as the actual DataFrame index rather than as a column, which was
what DictionaryIndex did. Making both use the DataFrame index
(since this is how index is stored in Index.metadata. Although in
Index.synonym_df, it's a column - in future we could consider making
these aligned