copy ontology_matching code from explosion arizona
Note: this DOESN'T pass the pre-commit checks - I have commited
this with --no-verify. That way, we have a clear trace of what
we've changed from the spacy code and why. This will be useful
in case they open source and we want to share our updates for
example. Following changes will also not pass all pre-commit
hooks as independent fixes/changes should get separate commits
for clarity