assert attributes are set to satisfy mypy
this satisfies mypy for the time being. Unclear if there might
be a better way to do it. The initialize method does give a default
value of None to parquet_files rather than requiring it for some
reason, so it would be possible for the assert to fail. However,
for now, I want to minimize how much I poke around inside the arizona
code.