bugfix for stopword removal
previously, removing "i" from the stopwords caused a KeyError
when running tests. This is because the Defaults object on the
pipeline is shared in-process, so if you instantiated multiple
StopWordRemover objects, the remove call on all but the first would
fail, as "i" was already removed. We don't want to mutate the
Defaults in a way that isn't obvious if we're using the pipeline
elsewhere, so use set difference and use the new created set.