remove count check in tests
this fails if we have a span with multiple entity types.
Explosion's code was set up so that this never happened, but
we've changed it as we want this in general. We could make
our tests more sophisticated to test for specific types
for each span only, but we are likely to rework what we're testing
here, so removing this check