fix use of requires_model_pack
I'd tried to attach this to a fixture so it was applied to all
tests using the fixture. This doesn't work. There are more
sophisticated ways to mark all all tests using a fixture, but it's
unclear if we want the complexity of these. We may not keep
these tests in their current arrangement anyway.