fix mypy issue - rewrite required
code was causing issues with mypy inferring types weirdly through
the zip and unpacking. I've re-written this in a way
that lets mypy do its thing and hopefully is more readable as well.
This may be dead code though - I opened issue #164 about this. I
tested equivalent with the old code using a quick hypothesis script
and setting max examples to 100000 as well as eyeballing output