simplify saving and loading of indices
save in a simpler location - this required moving and in some
cases regenerating cached versions of the indices. This relates
to the previous work simplifying the way indices are handled (e.g. the
old CachedIndexGroup and other concepts) - that work seemed to leave the
serialization of the indices in a state where it could be further
cleaned up in this commit.