run sapbert when there are no hits
Previously, we grouped hits by the parser and checked each group
has at least one exact match hit to decide if we wanted to skip
running sapbert. However, this meant that if there were no hits at
all (or no hits for a certain parser), we skipped sapbert (as
happened in the acceptance tests, or generally anywhere that we're
not running the dictionary linking step first).