make enums have auto values
No need to manually curate values, and use the name for
EquivalentIdAggregationStrategy rather than an arbitrary int.

LinkRanks needs to stay as ints (or be otherwise sortable) since we sort
based on the values in kb_disambiguation.py

Using auto() changes LinkRanks from starting at 1 to 0. This has no
impact though - we just use it for sorting so only the relative ordering
is important, not the exact values. The reason for auto() starting at
one is because Enum members evaluate to True, so it would be weird if
something with a value of 0 evaluated to True when 0 itself evaluates to
False in a boolean context.