use .get to avoid second call to synonym_db
previously, we were looking up a value in a dict where we'd just checked
the key was present but not fetched the value at the same time.

I was tempted to change synonym_db.get to behave like dict.get here
to further improve, but then all but 1 methods starting with 'get' on
the metadata and synonym databases use dict.__getattr (i.e. []), and
there's just one that behaves significantly differently in a
non-transparent way which would be bad. Changing all is too big a change
to be justified here.

Decided to rename 'term' to 'exact_match_term' otherwise mypy complains
further down as it get confused and thinks term could be None since it
would be used in two places, and one is Optional[SynonymTerm].