fix type hint
Iterable wasn't quite right, as we iterated over the terms, and then
tried to get the first item by iterating again. If this was a generator
expression or similar, the iterator would be exhausted and the
next(iter(synonyms)) call would raise an exception.

Changing to a Sequence, which was a bit stricter than necessary for the
previous code - we didn't really need to be able to index, only iterate
multiple times over the same iterable. However, also changed to index
for the first synonym rather than next(iter(synonyms)), as it's simpler
for a list or tuple (which seem the only realistic types to be passed
here), and then that does actually make Sequence pretty much exactly
right.

We could instead have messed around with itertools.chain so the
implementation worked even if we did get an Iterable that got exhausted
after iterating, but this adds complexity in the runtime code that isn't
justified when we only pass lists of synonyms so far.