use cached_property for BlackLister syns
This means they can just be calculated 'Just-in-Time' when we
call the blacklister for the first time. This is beneficial for two
reasons:
- we can instantiate blacklisters (and consequently
ExplosionNERStep) without having populated the synonym database, which
is tricky to guarantee has happened
- we never need to use memory for the synonym sets if we never call
the blacklisters. This currently happens at runtime, since we need the
blacklisters to instantiate the ExplosionNERStep, but we only use them
at spacy pipeline build time, rather than actually when we use them at
runtime.