pull out common synonym set building function
Could have put this on the BlackLister base class and actually inherited
(see later commit). However, this kindof means assuming all blacklisters
want to use this functionality. Really we want BlackLister just to
specify what a BlackLister needs to do and leave implementation up to
classes.

We could make a class that the existing blacklisters are all child
classes of that share functionality for building sets of synonyms (and
also using AnnotationLookups), but this seems like overkill for what we
save.