Make BlackLister class a Protocol
This was already really the case, just not explicitly, as none of the
blacklister classes inherited from this, we just used it as a type hint
elsewhere to check we were using valid methods.

A Protocol makes sense here, as we would really want anyone to be able
to come and write a new blacklister that 'just worked', and there's not
assumptions about how someone would want to implement any blacklister
that I can think of that's worth putting shared code on an actual
inherited base class.