add error handling and docstring
Add info to the error message so it's clearer what's actually going on.
This does add an entry to the stack trace which isn't ideal, but seems
clearer and less magic/hacky than other available options.

We also changes the type hint of synonym sources, so that we should only
get something that will print appropriately if an error does get raised.
This was already true in terms of what is actually used (always a
List[str]), just restricting the theoretical range of objects that can
be passed. In particular, if an Iterable that gets exhausted as it's
iterated on was passed, the error message wouldn't be very helpful.