create github external links
use the sphinx linkcode extension to link to the code in github from api
documentation. To get line number this requires some fiddling in the
linkcode_resolve function that this extension provides. Our version is
based off of pandas, in turn based off numpy, so we need to now include
their licenses as well (but these are both BSD 3-clause, so compatible).