configure furo theme's edit button
Now we get a link to edit on Github! Shortcoming - this is still present
for files which are autogenerated, which we don't really want (and would
be more of a problem if we stop having the built docs committed to the
repo). Kazu issue #251 to improve this, but this looks tricky. Having it
at all is still an improvement in my eyes.