generate docs for 'undocumented members'
This meant classes like OntologyMatcher didn't get any docs, even though
it's just that it doesn't have a docstring for the class. The __init__
and other methods have docstrings. In general, since this is intended as
a fairly complete reference, I think we want to err on the side of
generating documentation for too much rather than too little.

When we're a bit more rigorous on what we write docstrings for an how,
maybe we can switch this back off.