make sure printed info is logged
previously, if the test 'succeeds' (which this always will for
test_gold_standard_consistency, since there are no asserts, and the
prints within check_results_meet_threshold seem intended to go through
even when succesful overall), output is captured by pytest and not
reported. Default pytest behaviour is that output is only logged if the test fails.
Changing to always log for these print statements, using with
capsys.disabled().

some 'else' swapped for with capsys.disable() since the elses were
unnecessary - they were an else to the test failing, in which case
execution would stop anyway.