refactor AggregatedAccuracyResult
The way tasks related to items felt quite clunky, and we can now
recalculate less if properties are used multiple times.

Not quite a 'pure refactor' since the order of things returned for
tasks_for_fp and tasks_for_fn are different - they used to be
effectively grouped by the task, but now they're grouped by the item.
However, this isn't used anywhere in the checked in code.