fix broken docstrings
most problems around references to python code (esp. imported classes),
or numbered links in ReST, or whitespace.

This improves the docs quite a lot since often when the docs fail to
build properly, you get an outdated version of the file that doesn't
document the right documents/has an out of date docstring.