remove pipeline build logic from ExplosionNERStep
We can't easily build the pipeline at runtime anyway since the synonym
database needs to be populated, so we've decided to pull this out into a
separate script (which is the direction we want to go with the model
pack building process more generally anyway, so this is a step in the
right direction).

Removing the logic here makes the class easier to read, and means we
don't have parsers and all sorts hanging around where we don't need them
and they're just wasting memory.