remove blacklisting/exclusion code
We don't need this any more now we're moving to a whitelisting approach.
This also means we're only building a lowercase phrasematcher when we
build from parsers.

We still need the code to build from parsers so we can do an initial run
for new parsers/changed parsers so we can construct the whitelist.