handle cases where one phrasematcher is None
This is the norm now - standardly at runtime we will only have an
exact-match phrasematcher, and when introducing a new parser, we will
need to build a lowercase phrasematcher for it, to then curate synonyms
from to make the exact-match one later.

Note this will also speed up this step and probably reduce memory! Since
we only store and run a single phrasematcher. In principle, this should
be around twice as fast for this step - possibly even more since at
runtime it's the exact-match phrasematcher that runs, which should be
faster than the lowercase one since it doesn't have to run spacy's
lowercasing logic on the text to match against.