remove steps doc
This doesn't list all the steps, was erroring due to the rename of the
abbrevation step, and sphinx hated this file anyway because we were
calling autodoc here as well as elsewhere which it didn't like.

Currently more useful to browse through the 'steps' repo anyway, and in
the short-medium term we want some more narrative/tutorial docs that
explain stuf more fully anyway, so deleting since this is causing
problems without much benefit.