fix bug with multuple terms per SynonymTerm
we zipped together the SynonymTerms with the patters, derived from the
terms on those SynonymTerms, which meant if there were multiple terms on
a single SynonymTerm, the zip would misalign and cause problems.

The tests don't currently pick this up since they only test parsers with
a single term on each (generated) SynonymTerm. Test update to follow.