replace entity_class_to_ontology_map
Now we have an entity_class param for parsers, we can just collect the
information from there more easily, which avoids repetition and seems
like a more sensible place to store this information.

Note that this does change behaviour for string normalization of
synonyms for parsers where this wasn't previously specified and for
which the entity class causes the StringNormalizer to behave differently
to the default (e.g. Uberon for anatomy).

The new version also does mean we can't associate a parser/ontology with
multiple entity classes, which we could do before, but currently don't.
It's not clear to me that this is functionality we need. If we do in
future, we can always make 'entity_class' a list rather than a single
string (although it's unclear to me what the StringNormalizer would do
in that case).