Make parser name an instance attribute
change the name from a class attribute to an instance attribute. This is
necessary to allow creation of a custom RDFGraphParser purely through
config, which we want to support internal use cases where we don't want
to commit the parser to the open source code, and would prefer to avoid
the complexity of separate internal parser code. It also gives other
users of the code more flexibility in the same manner.