Clear cache when no GO parsers are missing from dbs
The Gene Ontology is large, so we don't really want to keep hold of a
parsed version in the cache longer than we need to. When all instances
that currently exist have populated the databases, clear the load_go
cache.

Note that if we later create new instances of GeneOntologyParser, we may
have to re-parse GO. This is expected behaviour - there is no way to
guarantee whether we will or won't call populate_databases again in the
future of a running program, so this seems like a good compromise.