reorder params to match order in parent method
I hadn't done this previously as there was just one that wasn't
'pass-through', so I wanted to highlight that by putting it at the end.
However, I now prefer keeping them where they were for the parent now
there's three of them, and they are fairly noticeable as a block.