remove unused depends_on
kazu is complex enough now that we can't sensibly programmatically
encode all the information about when it's reasonable to use a step
in the pipeline (especially while maintaining extensibility). As a
result, we now remove 'depends_on' and it is understood that in order
build a new/substantially different pipeline, you have to 'know what
you're doing'.