Re-use step error-handling code via decorators
Most Steps' __call__ method had error handling (and looping if
'per-doc') code that was substantially the same for every step.
Re-using/sharing this code via decorators to be applied to simpler
versions of the __call__ method on the steps.

One negative is that this makes it harder to read a step __call__ method
and tell exactly what's going on. One positive is that it makes it
clearer that Steps in general do their work via mutation of the input
documents.