apply step decorator to SethStep
Note that this does change behaviour - the try-except was previously at
a section level and now is at a doc level. As a result, if Seth fails
partway through a Document, it will no longer try any remaining sections
in that document.

This is probably less ideal, but doesn't seem worth the added
complexity/difference to the other iterating_steps. Additionally, it
seems to me like there are other steps that could behave the same way as
Seth, but previously didn't (as they still iterate section-by-section).
We could add a decorator like iterating_step but for section-level
iteration, and apply to all these steps. It seems like behaviour should
probably be consistent across steps here as much as possible.