add note about making Step.namespace a property
I tried this, and it was a bad idea! Noting why - we care more about
being able to get the namespace from the class without instantiating it
than we care about the niceness of it being a property/being able to
miss out the brackets/it matching how you access Entity.namespace.