Only try to remove base conf if it exists
This won't necessarily be present, and when it's not (e.g. trying to build just the merlin model pack and no others),
it previously threw a FileNotFoundError

Check if it exists instead of using ignore_errors param of shutil.rmtree as we don't want to ignore
e.g. permissions errors that actually are a problem.