fix skip messages
Motivation to change these was to be able to remove # noqa . However,
it also looked like the location of newlines was no intentional.
Changed this to what it seems was intended before.

Running pytest -v or pytest -vvv or pytest -rs before and after, there's
now a big improvement. Previously, with just pytest -v, nothing was
printed, as the skip messages started with a newline. With pytest -vv or
pytest -rs, the message was printed, but the starting newline break
looked weird. Now, we get a 'one-line summary' with pytest -v, and
pytest -vvv or pytest -rs prints the full message and the formatting
looks reasonable.