remove need for noqa and fix leading newline
get rid of the only # noqa comments left in the codebase - they're only
due to trailing whitespaces on the query lines which aren't necessary.
Also, escape the leading newline in the query, which doesn't seem
intentional.

These queries could probably be fiddled about with more if we wanted -
note that there's a lot of spurious indentation, as the indentation to
match the code's indentation still exists in the resultant string, so
the query basically has 3 level of spurious indentation. However, this
doesn't affect functionality, and makes the code easier to read than if
it broken indentation with the code. Leaving for now, as the main goal
was removing the # noqa , and there doesn't seem any obvious, very easy
'flypast' improvement (or if there is, it probably makes this
inconsistent with the way we do multi-line strings elsewhere).