Add in psutil dependency
We always depended on this, but got it through another dependency - but
now don't when I try installing it in a certain environment.

Note, exact version choice isn't done in a very principled way - just reading the
requests release notes, and choosing a release that seems 'pretty old,
but like not much has changed'.

Also, at time of writing, no testing of compatibility (though
incompatibility is unexpected).