handle undeclared tensorboard dependency
This was previously needed but not declared. Now adding as a dev
dependency, and changing how we import it so we don't try unless someone
actually tries to use the profiling functionality.

This makes sense in the dev dependencies, as it's likely people will
want to profile when developing kazu. In principle, we could have put it
in an extra 'extras_require' group, but this seems overcomplicated - if
people just want tensorboard, they can also just do 'pip install
tensorboard'.