auth fixes
First, we need add_id_header to have a 'if req_id is not None:' clause,
because otherwise, if we try to run the server with no authentication
middleware, there won't be an 'x-request-id' header (since this only
gets added by the authenticate method of JWTAuthenticationBackend, so
'req_id' will be None, which will then fail to encode in the response, which breaks
the requests.

Second, if we just fix that, the 'no auth middleware' version of the
server still gives errors if no auth is provided. Passing auto_error=False
to HTTPBearer fixes this, while still behaving correctly for the server
with auth middleware set up.