make CL parser __init__ match others
'name' param should be present on the parser, and required. This is
currently necessary for the config to work, which gives a 'name' keyword
argument to CLOntologyParser.__init__ , which currently throws an
exception since CLOntologyParser.__init__ doesn't have a 'name'
argument, as it 'hard-codes' passing "CL" to OntologyParser.__init__

I could have fixed this by removing that 'name' keyword argument in the
config, but that would mean we couldn't interpolate the parser_name for
the additional_synonyms parameter construction, which is why a lot of
the __init__'s of parsers changed in the first place. So sticking with
consistency with the other 'child' parsers with custom __init__s.