avoid add_id_header completely when auth isn't used
Previously, we made add_id_header do nothing if the relevant header
wasn't present (when the auth middleware isn't present), but this still
costs us computation as the middleware still processes all requests.
Instead, use the config along with the auth middleware, so it can be
completely removed when not using jwt auth.