pass algorithm as list as expected
mypy flagged that this should be a list, which seems true. It only
'happens' to work because in jwt/api_jws.py , PyJWS._verify_signature
checks if the alogrithm in the header is in the valid algorithms by
doing 'alg not in algorithms' and the str __contains__ methods will
return true with the same string, so it 'kind of' works anyway. This is
an implementation detail though, using the declared type is more robust
to future modification.