Use namespace classmethod where possible for StrategyRunner
Safer than raw strings across refactors as will fail loudly. Note though
that the same config file has to use the 'raw' strings as keys as part
of the symbolic_strategies config etc. There's no easy way around this
with hydra, but we can make a change to have assertions to compare these
namespaces with the ones in the processing order to protect against
problems here.