add assertions to make namespace use safer
As in previous commit's description, we're using raw strings for some
(but not all) of the config of StrategyRunner. By comparing the
namespaces in the processing order to those in symbolic_strategies and
non_symbolic_strategies, we prevent errors in and of itself, but this is
particularly of note since it 'fill in' the lack of protection against
refactors of the 'raw' strings of the namespaces, since we're now
comparing them against namespace strings that come from the classmethod
calls. So now if there's an issue, it should be very likely to fail
loudly and early, where before there were several opportunities for
silent or late mis-configuration.