fix check for empty sort order: None to empty list
Previously, we were checking if it was None, but it never is from the
type signature - so the 'empty' value is an empty list, not None. Also
reverse the order of the clauses as it seems easier to read - less time
to hold the 'if' in your brain while you're reading the code this way
round.