document dataclass attributes as attributes rather than like function
Previously, we get a nasty generated html file, where the
BuildConfiguration class gets some documented params, but then there's
also the generated __init__ with the same params, but not documented,
and then a list of attributes matching the same params, without docs.

This improves on this slightly, by removing the params in the
BuildConfiguration class itself, and putting them on the documented
attributes. However, we still have the __init__ docs.