update github action versions from marketplace
We are currently getting a couple of sets of warnings in our actions:

1. saying that Node.js 12 actions are deprecated, and to upgrade actions
   to use Node.js 16:
   https://github.blog/changelog/2022-09-22-github-actions-all-actions-will-begin-running-on-node16-instead-of-node12/
2. saying that the `save-state` command is deprecated and will be disabled soon:
   https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/

Both of these result from 'old' version of actions pulled in from the
actions marketplace, tha use Node.js 12, or use the 'save-state' command
within this action. Upgrading to the most recent major versions to
minimize these warnings. I had a quick skim through the release notes of
the actions between the version we were previously on, and the new
version, and there wasn't anything that looked worrying to me.