use torch.softmax for consistency
Replace a call to torch.nn.functional.softmax with its alias (according
to the docs) torch.softmax, so it's consistent with the torch.sigmoid
call.

This does make it inconsistent with the use of torch.nn.functional.softmax
in kazu/modelling/distillation/models.py , but that usage is consistent
with the use of torch.nn.functional.log_softmax in the same file
(torch.log_softmax weirdly does exist, but isn't documented in the torch
docs at all, and code/behaviour isn't clear enough for me to trust it).
I think it's better to have each file be consisent.