move to hatch for build and versioning
Hatch is managed by PyPA, and is in the official tutorial:
https://packaging.python.org/en/latest/tutorials/packaging-projects/

It makes versioning nicer because:
a.) We only have version defined in a single place, and a single config
file pointing to that (as opposed to it being defined by the bumpversion
config, written out in kazu/__init__.py, but also pointed to by the
setuptools build config, it's now just that the 'master value' is in
kazu/__init__.py, and we point to that from hatch's build config).
b.) We replace multiple tools with a single, well suported one
(bumpversion and setuptools with hatch). This is even more true since
(when we fix the release job for the new build setup) we'll be able to
replace twine with hatch publish as well.