make pytest testpaths more specific
pytest has fewer files/directories to search this way, so it should save
work. This is supported by the output of 'pytest --collect-only' before
and after the change.

When doing that, the same number of tests are found, but running 3 times
for each, this change does seem to speed up test collection a little
(not putting numbers because it's such a small sample and I wasn't
taking care not to run other stuff on my laptop at the same time, so I
don't think the numbers would be meaningful).