update pip caching key to use pyproject.toml
This previously used setup.py, which was where the dependencies were
defined (which is what we want to hash on really). Dependencies are now
in pyproject.toml and setup.py doesn't exist, so use pyproject.toml.

This possibly will get invalidated incorrectly more with pyproject.toml,
since there's more stuff in there (like black and pytest config), but we
acccepted this happening to some degree before with setup.py anyway, and
we don't expect pyproject.toml to change too often anyway. Maybe we
could do something more sophisticated, but it doesn't seem worth the
added complexity for the benefit at this time.