move mypy flags to pyproject.toml
This means if you invoke 'mypy' as a standalone command, it will now
behave the same as it does if you do 'pre-commit run -a mypy'. Saves you
remembering the relevant command line flags.

It also open a path for us doing more strict mypy checking
incrementally, as we can be more strict for a particular file/module.