fix hydra env_set interpolation issue
Previously, we were trying to interpolation into a bit of the config in
the 'main' config from within the 'hydra' config. Apparently you can't
do this. Workaround using a global package to still have all the cache
size config in a single file for discoverability and ease of changing
all.