log when user is unauthorized as well as raising
This matches the pattern below for when there's an invalid jwt token,
and means we can more easily see if we're receiving requests from
unauthorized users.

Note that we don't use the logging built-in formatting here, because the
message is shared with the error, which we know we're going to raise and
therefore need.