Add in mention confidence
This will let us say that some Entity matches, we're not confident the
match is good (e.g. symbolic from Explosion), whereas others we are.

Add this as an enum rather than a boolean flag like 'putative' in case
we want to expand in future to 'middle' positions like 'PROBABLE'.