Refactor CuratedTerm / Curation dataclass
Make this both easier to understand, and give us more flexibility, for
example enabling us dropping 'COX' as a synonym for specific gene
parsers ('opentargets_target') and not others ('gene_family'), or even
for specific IDs.

It's also simple from a modelling perspective to allow a hierarchy here
so that we can do things like 'drop it for all indication parsers apart
from ...' which will make managing these kinds of overrides simpler.

Co-authored-by: Richard Jackson <richard.jackson4@astrazeneca.com>