Add Trove classifiers for PyPI, remove license field
These give metadata about the project that will be shown in PyPI, and
categorize the project.

I chose the classifiers based on skimming the full list here, and adding
anything that seemed relevant:
https://pypi.org/classifiers/

Now we have the license classifier, we don't need the license field
populated. The 'license' field is defined as:

"Text indicating the license covering the distribution where the license
is not a selection from the “License” Trove classifiers. "

https://packaging.python.org/en/latest/specifications/core-metadata/#license

And the packaging tutorial specifies a license classifer but not the
'license' field:

https://packaging.python.org/en/latest/tutorials/packaging-projects/