simplify Section._sentence_spans
Previously we were using a dict with values we didn't use, in order to
get the behaviour of an immutable, ordered set (i.e. without
duplicates). We don't actually expect duplicates though, so switch to a
tuple, and add an assert to check for duplicates.

I considered that for Section.sentence_spans, maybe we could now just
return specifically a tuple rather than an iterable of CharSpan . tuple
is an iterable, so this won't break anyone's code, but lets users know
that they can do things like index into the result, and iterate without
'exhausting' the iterable. However, mypy can't handle the case of the
type of the setter not being the same as that of the getter  - see
https://github.com/python/mypy/issues/3004

If mypy fix this in future, we can upgrade mypy and narrow the type hint
here. Otherwise, I don't think it's worth the pain of having #  type:
ignore comments for users.