Make CharSpan frozen
This also means we get a generated hash function for free which is
identical to the current explicitly defined one, so we can delete that.

Unfortunately, we can't make 'order=True' as the behaviour wouldn't be
the same - we need to compare the start for less than, but the end for
greater than, which we couldn't do with 'order=True'.