use absolute logo link for README, to work in PyPI
At the moment, this only works on GitHub, where the file is present to
link to. For PyPI, we need an absolute link.

One downside here - in principle, this means that if we update the logo,
old releases might show the new logo. If we moved the logo, old releases
might have a broken logo link.

Given we're unlikely to move the logo, that doesn't seem too much of an
issue, and if we update the logo, it doesn't seem like really an issue
to me that the old releases have the new logo.