add test to keep tokenization infixes updates
Check that we haven't drifted from the spacy tokenization and are losing
improvements to their rules.

Possibly a bit cheeky since this isn't really a 'test' of our code
working, but it seems like an easy way of getting the desired behaviour
of alerting kazu devs if our tokenization rules drift out of alignment.