remove web config
This started as a simpler version, but now we've moved most of the
unwanted complexity out of the default model pack and into custom
internal model packs, so the default pipeline is suitable now.

Also beneficial for users who try out the pipeline using the frontend,
but then use the default pipeline locally and wonder why it's not
behaving the same.