docs: fix whitespace for nested lists
This was previously causing Sphinx to decide these were blockquotes and
therefore introduce a html blockquote node with related styling. This
fixes that.

The styling maybe isn't super cool for the nested lists, but I think we
should basically get the html right, and if the styling isn't ideal,
fiddle with css to make it prettier if we care enough.