Make Sphinx doc links point to opensource repo
I'd tried to be clever and use the git config, assuming that when we
went open source, we would just have a single open-source git remote.
However, that assumption was wrong, and has meant we've kept generating
links in our sphinx docs that external users will not be able to access.

This commit stops trying to be clever and just 'hard-codes' the open
source URL - we don't expect this to change in future, and if it does,
it will be easier to find and update by searching for the string
AstraZeneca/KAZU.