Fix bibtex in docstrings
Sphinx was trying to interpret this is ReST, so then throwing errors and
failing to render nicely. Instead, I've added a text version of the
citation (either as given by the authors, or as produced when actually
rendering the bibtex citation), make the bibtex a bibtex codeblock, and
added a collapsible element around it to make the output cleaner.

Related minor changes:

- Abbreviation_detector module level docstring wasn't rendering because it
  needed to be before the imports
- make the license for Gilda in a collapsible element as well for
  brevity