Call ray init and serve.start with hydra.utils.call
This gives more control to the end user via configuration in case new
options need to be passed to init and call.

It also fixes the fact that we weren't actually using the port in the
config.

It also allows a simpler way of specifying that auth is required by the
middlewares, rather than trying to figure that out based on the class of
middlewares used.