docs: fix broken Sphinx doctest
This broken when we moved the main code inside a hydra.main decorated
function and an if __name__ == "__main__" code block. When Sphinx runs
this, __name__ ISN'T "__main__", it's "builtins", so the function
doesn't run, and the test fails, because there's no output.

Adding a hidden codeblock below that manually does run kazu_test, and
supplies the config (otherwise hydra tried to use argparse, which causes
Sphinx to throw an error).