ci: ensure sphinx doctest gets run
Previously, this was being skipped because it didn't have
KAZU_MODEL_PACK set when we were running it.

Since it depends on the model pack, it should really use the newly built
version anyway to make sure that works. It need to be inside 'Build
model pack caches and test' based on how that works currently, since it
deletes the virtualenv and model pack build path at the end of that
step.