fix: hydra defaults for acceptance tests
When running the acceptance tests standalone as the top-level function,
update the default hydra config path and config name. The name is
particular currently seems 'wrong' as the main hydra config is
'config.yaml'. A config path of '.' seems more useful as well that then
current '../../'.

Note that typically, we call this function, so this wasn't 'broken' per
say, especially since the config path and config names are overridable.
However, these seem like more appropriate defaults.