add top-level directory to zip files in build process
We currently have these on the public model packs, but the current build
process doesn't have them in the same way (although unzipping through
MacOS Finder obscures this). Adding them in so we match the model pack,
plus so that if someone unzips it, they get a model pack folder with the
version, which is useful for the KAZU_MODEL_PACK env variable they will
need.