docs: make sphinx flag failed links
By doing this, if we add a link that's actually broken, we'll know about
it now (from the integration tests failing), whereas we didn't before.

Having to ignore our existing failed links since there's a lot, and some
of them look fairly difficult to resolve when it's type hints that are
being attempted to be linked.