docs: remove redundant rtype annotations
This allowed us to partially resolve the problem that the previous
commit's new sphinx extension solves, but it doesn't solve all cases.
For example, there was no way of manually altering the docs like this
for dataclass properties.

Now we have the new custom sphinx extension, the :rtype: bits here are
redundant, as the links populate correctly without them.