docs: custom title for xml.dom.minidom types
Previously, we had it all working automatically, but the link title was
a bit cryptic/confusing. Use manual :rtype: or :type: sphinx field list
items so we can control the link title.

Doing that also means we can control the link target from there, so this
does everything we need for sphinx to behave nicely here. As a result,
we can get rid of the 'tuple' option for the sphinx custom extension for
intersphinx overrides, which complicated the config, but failed to do
everything we wanted here (because it didn't as-is have the ability to
set the title, and seemed hard to add this in with massively
overcomplicating).