Fix inherited docstring by writing new one
There was a failed sphinx link for 'Module' which came from one of the
base classes, where Sphinx didn't know where it was coming from.

Adds a new docstring to 'cover up' this problematic one, where we can
link explicitly to the torch.nn.Module class that was causing problems.
This also takes away a comment about 'subclasses should override' which
was relevant for the base class but not our class, so this is an
improvement in general.